Moving call to chatroom
