+++ samples/apps/copilot-chat-app/webapp/src/components/chat/ChatInput.tsx
@@ -14,7 +14,7 @@ import { useAppDispatch } from '../../redux/app/hooks';
 import { addAlert } from '../../redux/features/app/appSlice';
 import { useSKSpeechService } from './../../libs/semantic-kernel/useSKSpeech';
 import { TypingIndicatorRenderer } from './typing-indicator/TypingIndicatorRenderer';
-import { useSKMultiUserChat } from './../../libs/semantic-kernel/useSKMultiUserChat';
+// import { useSKMultiUserChat } from './../../libs/semantic-kernel/useSKMultiUserChat';
 
 const log = debug(Constants.debug.root).extend('chat-input');
 
@@ -71,18 +71,18 @@ export const ChatInput: React.FC<ChatInputProps> = (props) => {
     const [recognizer, setRecognizer] = React.useState<speechSdk.SpeechRecognizer>();
     const [isListening, setIsListening] = React.useState(false);
     const speechService = useSKSpeechService(process.env.REACT_APP_BACKEND_URI as string);
-    const chatRelay = useSKMultiUserChat(process.env.REACT_APP_BACKEND_URI as string);
+    // const chatRelay = useSKMultiUserChat(process.env.REACT_APP_BACKEND_URI as string);
     const [documentImporting, SetDocumentImporting] = React.useState(false);
     const documentImportService = useDocumentImportService(process.env.REACT_APP_BACKEND_URI as string);
     const documentFileRef = useRef<HTMLInputElement | null>(null);
 
-    React.useEffect(() => {
-        // This code will run once, when the component is mounted
-        console.log('SignalR setup called');
-        chatRelay.setupSignalRConnectionToChatHub();
-    // Disabling warning so that we can use empty dependency array to invoke this setup call just once 
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, []);
+    // React.useEffect(() => {
+    //     // This code will run once, when the component is mounted
+    //     console.log('SignalR setup called');
+    //     chatRelay.setupSignalRConnectionToChatHub();
+    // // Disabling warning so that we can use empty dependency array to invoke this setup call just once 
+    // // eslint-disable-next-line react-hooks/exhaustive-deps
+    // }, []);
 
     React.useEffect(() => {
         if (recognizer) return;
